Move astropy_support to units.py per reviewer feedback
- Moved astropy_support function from separate contexts.py to units.py
- Removed contexts.py module as requested by reviewer
- Updated __init__.py imports
